Add reproducer
I'd like to be able to factor out the repeated code I use to open a file
or stdin, but I'm not sure how I can make a function return a generic
type that is not based on the type of the input, but is instead based on
the value of an input. Here's a minimal reproducer I can point to when
asking for help.